requirements.txt
# ============================================================
# REQUIREMENTS POUR LE PROJET DE RECHERCHE SÉMANTIQUE
# ============================================================

# Base de données
psycopg2-binary==2.9.9    # Driver PostgreSQL pour Python
pgvector==0.2.5            # Extension pour vecteurs dans PostgreSQL

# Machine Learning / Embeddings
sentence-transformers==2.2.2  # Modèles d'embeddings
torch>=2.0.0                  # PyTorch (requis par sentence-transformers)
transformers>=4.30.0          # Transformers de Hugging Face

# Data Processing
pandas>=2.0.0              # Manipulation de données
numpy>=1.24.0              # Calculs numériques
scikit-learn>=1.3.0        # Métriques et preprocessing

# Dataset
datasets>=2.14.0           # Hugging Face datasets

# Visualisation et Utilités
tqdm>=4.65.0               # Barres de progression
matplotlib>=3.7.0          # Graphiques (optionnel)
seaborn>=0.12.0            # Graphiques améliorés (optionnel)

# Web Interface (optionnel pour plus tard)
streamlit>=1.28.0          # Interface web interactive

# Logging et Debugging
python-dotenv>=1.0.0       # Variables d'environnement
colorama>=0.4.6            # Couleurs dans le terminal (Windows)

# Testing (optionnel)
pytest>=7.4.0              # Tests unitaires
streamlit
